fix: package YARA guard without path dependency
