+++ src/security/mod.rs
@@ -26,6 +26,7 @@
 
 pub mod fetch_yara;
 pub mod ingestion_guard;
+pub mod yara_engine;
 
 pub use fetch_yara::{guard_fetch_output, guard_fetch_output_with_config};
 pub use ingestion_guard::{DetectionReport, DirectiveKind, Sample, Severity, detect, sanitize};
